fix: database connection in unit tests
